requirements.txt
appdirs==1.4.4
apt-xapian-index==0.49
argon2-cffi==20.1.0
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.1
bidict==0.21.2
bleach==3.2.1
blinker==1.4
blosc==1.8.1
Bottleneck==1.2.1
Brlapi==0.7.0
capstone==4.0.1
certifi==2020.4.5.1
cffi==1.14.4
chardet==3.0.4
chrome-gnome-shell==0.0.0
clarifai-grpc==7.2.0
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.3
command-not-found==0.3
cryptography==3.0
cupshelpers==1.0
cycler==0.10.0
dask==2.11.0+dfsg
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.5.0
distro-info===0.23ubuntu1
dnspython==1.16.0
ecdsa==0.15
entrypoints==0.3
enum34==1.1.10
et-xmlfile==1.0.1
eventlet==0.26.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==4.2.1
fsspec==0.6.1
future==0.18.2
googleapis-common-protos==1.52.0
gpg===1.14.0-unknown
greenlet==0.4.15
grpcio==1.35.0
gyp==0.1
h5py==2.10.0
h5py.-debian-h5py-serial==2.10.0
hidpidaemon==18.4.6
html5lib==1.1
httplib2==0.18.1
idna==2.10
imageio==2.4.1
importlib-metadata==1.6.0
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
itsdangerous==1.1.0
jdcal==1.0
jedi==0.18.0
jeepney==0.4.3
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kernelstub==3.1.4
keyring==21.3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
llvmlite==0.33.0
locket==0.2.0
louis==3.14.0
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
msgpack==0.6.2
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netifaces==0.10.4
networkx==2.5
nose==1.3.7
notebook==6.1.6
numba==0.50.1
numexpr==2.7.1
numpy==1.18.4
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.3
packaging==20.8
pandas==1.0.5
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.1
partd==1.0.0
patsy==0.5.1
pbr==5.5.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.14.3
pluggy==0.13.1
pop-transition==1.1.2
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.12.3
psutil==5.8.0
ptyprocess==0.7.0
pwntools==4.3.1
py==1.10.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.9.7
pycuda==2018.1.1
pycups==2.0.1
pydbus==0.6.0
pyelftools==0.27
Pygments==2.7.3
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pynvim==0.4.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypillowfight==0.3.0
PyQt5==5.15.0
pyRFC3339==1.1
pyrsistent==0.17.3
pyserial==3.5
PySocks==1.7.1
pytest==6.2.1
python-apt==2.1.3+ubuntu1.3
python-dateutil==2.8.1
python-debian==0.1.37
python-engineio==4.0.0
python-ptrace==0.9.7
python-socketio==4.4.0
python-xlib==0.27
pytools==2021.1
pytz==2020.1
PyWavelets==1.1.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
repolib==1.5.2
repoman==1.4.0
reportlab==3.5.47
requests==2.23.0
requests-unixsocket==0.2.0
retrying==1.3.3
ROPGadget==6.4
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
screen-resolution-extra==0.0.0
seaborn==0.10.1
SecretStorage==3.1.2
Send2Trash==1.5.0
sessioninstaller==0.0.0
simplejson==3.17.0
sip==4.19.24
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.0.1
ssh-import-id==5.10
systemd-python==234
tables==3.6.1
terminado==0.9.2
testpath==0.4.4
testresources==2.0.1
threadpoolctl==2.1.0
tifffile==2021.1.11
toml==0.10.2
toolz==0.9.0
torch==1.7.1
torchvision==0.8.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
ubuntu-advantage-tools==24.4
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==4.0.2
unicorn==1.0.2rc1
urllib3==1.25.9
usb-creator==0.3.7
virtualenv==20.0.29+ds
wadllib==1.3.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
youtube-dl==2020.9.14
zipp==1.0.0
